# api/01-openapi-spec.yaml
openapi: 3.0.3
info:
  title: Ride Circle API
  description: |
    Comprehensive API for the Ride Circle peer-to-peer car rental platform.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Public endpoints: 100 requests/minute
    - Authenticated endpoints: 1000 requests/minute
    - Admin endpoints: 5000 requests/minute
    
    ## Multi-tenancy
    All requests must include the tenant context via subdomain (mumbai.ridecircle.in) 
    or X-Tenant-Code header.
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format with additional context.
    
  version: 1.0.0
  contact:
    name: Ride Circle API Support
    email: api-support@ridecircle.in
  license:
    name: Proprietary
servers:
  - url: https://api.ridecircle.in/v1
    description: Production server
  - url: https://staging-api.ridecircle.in/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Register a new user with email/phone verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, phone, password, firstName, lastName, role]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                phone:
                  type: string
                  pattern: '^\+91[6-9]\d{9}$'
                  example: "+919876543210"
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123!"
                firstName:
                  type: string
                  minLength: 2
                  example: "John"
                lastName:
                  type: string
                  minLength: 2
                  example: "Doe"
                role:
                  type: string
                  enum: [customer, host]
                  example: "customer"
                referralCode:
                  type: string
                  example: "REF123"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user with email/phone and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier:
                  type: string
                  description: Email or phone number
                  example: "user@example.com"
                password:
                  type: string
                  example: "SecurePass123!"
                deviceInfo:
                  type: object
                  properties:
                    deviceId:
                      type: string
                    platform:
                      type: string
                      enum: [web, ios, android]
                    version:
                      type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Account locked due to failed attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/verify-otp:
    post:
      tags: [Authentication]
      summary: Verify OTP
      description: Verify OTP for phone/email verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contact, otp, purpose]
              properties:
                contact:
                  type: string
                  example: "+919876543210"
                otp:
                  type: string
                  pattern: '^\d{6}$'
                  example: "123456"
                purpose:
                  type: string
                  enum: [registration, login, password_reset]
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /admin/analytics/dashboard:
    get:
      tags: [Admin]
      summary: Get dashboard analytics
      description: Get key metrics for admin dashboard
      security:
        - BearerAuth: [admin]
      parameters:
        - name: period
          in: query
          description: Time period for analytics
          schema:
            type: string
            enum: [today, week, month, quarter, year]
            default: month
      responses:
        '200':
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAnalytics'
        '403':
          $ref: '#/components/responses/Forbidden'

  # =====================================================
  # UTILITY ENDPOINTS
  # =====================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        errors:
          type: array
          items:
            type: string
        code:
          type: string
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
            tokens:
              $ref: '#/components/schemas/TokenResponse'
        message:
          type: string
          example: "Authentication successful"
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expiresIn:
          type: integer
          example: 900
        tokenType:
          type: string
          example: "Bearer"

    # User schemas
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
          enum: [customer, host, admin]
        status:
          type: string
          enum: [pending, active, suspended, banned]
        firstName:
          type: string
        lastName:
          type: string
        profileImageUrl:
          type: string
          format: uri
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        kycStatus:
          type: string
          enum: [pending, in_review, approved, rejected]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminAnalytics:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            overview:
              type: object
              properties:
                totalUsers:
                  type: integer
                totalVehicles:
                  type: integer
                totalBookings:
                  type: integer
                totalRevenue:
                  type: number
                activeBookings:
                  type: integer
            growth:
              type: object
              properties:
                userGrowth:
                  type: number
                vehicleGrowth:
                  type: number
                bookingGrowth:
                  type: number
                revenueGrowth:
                  type: number
            charts:
              type: object
              properties:
                bookingsByDay:
                  type: array
                  items:
                    type: object
                    properties:
                      date:
                        type: string
                        format: date
                      count:
                        type: integer
                revenueByMonth:
                  type: array
                  items:
                    type: object
                    properties:
                      month:
                        type: string
                      revenue:
                        type: number
                topVehicles:
                  type: array
                  items:
                    type: object
                    properties:
                      vehicleId:
                        type: string
                      make:
                        type: string
                      model:
                        type: string
                      bookings:
                        type: integer
                      revenue:
                        type: number

    # Utility schemas
...
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Validation failed"
            errors: ["Email is required", "Phone number is invalid"]
            code: "VALIDATION_ERROR"
            timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            code: "UNAUTHORIZED"
            timestamp: "2024-01-15T10:30:00Z"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Access denied"
            code: "FORBIDDEN"
            timestamp: "2024-01-15T10:30:00Z"
...
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Email already registered"
            code: "CONFLICT"
            timestamp: "2024-01-15T10:30:00Z"
...