# app/api/v1/admin/analytics/dashboard/route.ts
import { type NextRequest, NextResponse } from "next/server"

export async function GET(request: NextRequest) {
  try {
    // TODO: Verify JWT token and check admin role
    const authHeader = request.headers.get("authorization")
    if (!authHeader) {
      return NextResponse.json(
        {
          success: false,
          message: "Authentication required",
        },
        { status: 401 },
      )
    }

    // TODO: Get analytics data from database
    // Mock analytics data for demonstration
    const analyticsData = {
      overview: {
        totalUsers: 15420,
        totalHosts: 3240,
        totalCustomers: 12180,
        totalVehicles: 8950,
        activeBookings: 234,
        totalRevenue: 2450000, // in INR
        monthlyGrowth: {
          users: 12.5,
          bookings: 18.3,
          revenue: 22.1,
        },
      },
      bookings: {
        today: 45,
        thisWeek: 312,
        thisMonth: 1240,
        totalCompleted: 18450,
        averageBookingValue: 3200,
        cancellationRate: 8.5,
        statusBreakdown: {
          pending: 23,
          confirmed: 156,
          active: 34,
          completed: 89,
          cancelled: 12,
        },
      },
      revenue: {
        today: 144000,
        thisWeek: 998400,
        thisMonth: 3960000,
        platformFee: 198000, // 5% of monthly revenue
        hostPayouts: 3564000, // 90% of monthly revenue
        pendingPayouts: 234000,
        averageCommission: 5.2,
      },
      vehicles: {
        totalListed: 8950,
        activeListings: 7820,
        pendingApproval: 145,
        suspended: 85,
        averageUtilization: 68.5,
        topPerformingCategories: [
          { category: "Hatchback", count: 3240, utilization: 72.3 },
          { category: "Sedan", count: 2890, utilization: 69.8 },
          { category: "SUV", count: 1820, utilization: 65.2 },
        ],
      },
      users: {
        newRegistrations: {
          today: 23,
          thisWeek: 156,
          thisMonth: 640,
        },
        kycStatus: {
          pending: 234,
          approved: 14890,
          rejected: 296,
        },
        userActivity: {
          activeUsers: 8920,
          averageSessionDuration: 12.5, // minutes
          bounceRate: 23.4,
        },
      },
      geography: {
        topCities: [
          { city: "Mumbai", users: 4230, bookings: 890, revenue: 890000 },
          { city: "Delhi", users: 3890, bookings: 780, revenue: 780000 },
          { city: "Bangalore", users: 3240, bookings: 650, revenue: 650000 },
          { city: "Pune", users: 2340, bookings: 470, revenue: 470000 },
          { city: "Chennai", users: 1720, bookings: 340, revenue: 340000 },
        ],
        expansionOpportunities: [
          "Hyderabad - High demand, low supply",
          "Kolkata - Growing market potential",
          "Ahmedabad - Emerging opportunity",
        ],
      },
      incidents: {
        totalReported: 89,
        resolved: 76,
        pending: 13,
        averageResolutionTime: 2.3, // days
        typeBreakdown: {
          damage: 34,
          accident: 12,
          theft: 3,
          breakdown: 28,
          traffic_violation: 12,
        },
      },
      performance: {
        systemUptime: 99.8,
        averageResponseTime: 245, // milliseconds
        errorRate: 0.12,
        activeConnections: 1240,
        queueHealth: {
          notifications: { pending: 23, processing: 5, failed: 1 },
          payments: { pending: 12, processing: 3, failed: 0 },
          kyc: { pending: 45, processing: 8, failed: 2 },
        },
      },
    }

    console.log("[v0] Admin analytics retrieved:", { timestamp: new Date().toISOString() })

    return NextResponse.json({
      success: true,
      message: "Analytics data retrieved successfully",
      data: analyticsData,
      metadata: {
        generatedAt: new Date().toISOString(),
        dataFreshness: "Real-time",
        reportingPeriod: "Current month",
      },
    })
  } catch (error) {
    console.error("[v0] Admin analytics error:", error)
    return NextResponse.json(
      {
        success: false,
        message: "Failed to retrieve analytics data",
      },
      { status: 500 },
    )
  }
}